+++ lib/msal-angular/src/msal.broadcast.service.spec.ts
@@ -1,5 +1,6 @@
 import { TestBed } from "@angular/core/testing";
 import {
+  EventHandler,
   EventType,
   InteractionStatus,
   InteractionType,
@@ -19,8 +20,17 @@ const msalInstance = new PublicClientApplication({
   },
 });
 
+let eventHandler;
+const eventSpy = jasmine.createSpy("addEventCallback");
+msalInstance.addEventCallback = eventSpy;
+
 function initializeMsal(providers: any[] = []) {
   TestBed.resetTestingModule();
+
+  eventHandler = new EventHandler();
+  eventSpy.and.callFake((callback, types) => {
+    return eventHandler.addEventCallback(callback, types);
+  });
 
   TestBed.configureTestingModule({
     imports: [
@@ -68,10 +78,7 @@ describe("MsalBroadcastService", () => {
       }
     });
 
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.LOGIN_START, InteractionType.Popup);
+    eventHandler.emitEvent(EventType.LOGIN_START, InteractionType.Popup);
   });
 
   it("broadcasts previous events if MsalBroadcastConfig set and eventsToReplay is greater than 0", (done) => {
@@ -115,10 +122,7 @@ describe("MsalBroadcastService", () => {
       }
     });
 
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.LOGIN_START, InteractionType.Redirect);
+    eventHandler.emitEvent(EventType.LOGIN_START, InteractionType.Redirect);
 
     subscription.unsubscribe();
 
@@ -161,14 +165,14 @@ describe("MsalBroadcastService", () => {
       }
     );
 
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.HANDLE_REDIRECT_START, InteractionType.Redirect);
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.HANDLE_REDIRECT_END, InteractionType.Redirect);
+    eventHandler.emitEvent(
+      EventType.HANDLE_REDIRECT_START,
+      InteractionType.Redirect
+    );
+    eventHandler.emitEvent(
+      EventType.HANDLE_REDIRECT_END,
+      InteractionType.Redirect
+    );
 
     newSubscription.unsubscribe();
   });
@@ -220,14 +224,11 @@ describe("MsalBroadcastService", () => {
       }
     });
 
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.INITIALIZE_START, InteractionType.Redirect);
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.LOGIN_START, InteractionType.Redirect);
+    eventHandler.emitEvent(
+      EventType.INITIALIZE_START,
+      InteractionType.Redirect
+    );
+    eventHandler.emitEvent(EventType.LOGIN_START, InteractionType.Redirect);
 
     subscription.unsubscribe();
 
@@ -270,14 +271,14 @@ describe("MsalBroadcastService", () => {
       }
     );
 
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.HANDLE_REDIRECT_START, InteractionType.Redirect);
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.HANDLE_REDIRECT_END, InteractionType.Redirect);
+    eventHandler.emitEvent(
+      EventType.HANDLE_REDIRECT_START,
+      InteractionType.Redirect
+    );
+    eventHandler.emitEvent(
+      EventType.HANDLE_REDIRECT_END,
+      InteractionType.Redirect
+    );
 
     newSubscription.unsubscribe();
   });
@@ -314,10 +315,7 @@ describe("MsalBroadcastService", () => {
       }
     });
 
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.LOGIN_START, InteractionType.Redirect);
+    eventHandler.emitEvent(EventType.LOGIN_START, InteractionType.Redirect);
 
     subscription.unsubscribe();
 
@@ -354,14 +352,14 @@ describe("MsalBroadcastService", () => {
       }
     );
 
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.HANDLE_REDIRECT_START, InteractionType.Redirect);
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.HANDLE_REDIRECT_END, InteractionType.Redirect);
+    eventHandler.emitEvent(
+      EventType.HANDLE_REDIRECT_START,
+      InteractionType.Redirect
+    );
+    eventHandler.emitEvent(
+      EventType.HANDLE_REDIRECT_END,
+      InteractionType.Redirect
+    );
 
     newSubscription.unsubscribe();
   });
@@ -382,14 +380,11 @@ describe("MsalBroadcastService", () => {
       }
     });
 
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.HANDLE_REDIRECT_START, InteractionType.Redirect);
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.LOGIN_SUCCESS, InteractionType.Redirect);
+    eventHandler.emitEvent(
+      EventType.HANDLE_REDIRECT_START,
+      InteractionType.Redirect
+    );
+    eventHandler.emitEvent(EventType.LOGIN_SUCCESS, InteractionType.Redirect);
   });
 
   it("HANDLE_REDIRECT_END event sets inProgress to None if handleRedirect is in progress", (done) => {
@@ -409,18 +404,15 @@ describe("MsalBroadcastService", () => {
       }
     });
 
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.HANDLE_REDIRECT_START, InteractionType.Redirect);
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.LOGIN_SUCCESS, InteractionType.Redirect);
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.HANDLE_REDIRECT_END, InteractionType.Redirect);
+    eventHandler.emitEvent(
+      EventType.HANDLE_REDIRECT_START,
+      InteractionType.Redirect
+    );
+    eventHandler.emitEvent(EventType.LOGIN_SUCCESS, InteractionType.Redirect);
+    eventHandler.emitEvent(
+      EventType.HANDLE_REDIRECT_END,
+      InteractionType.Redirect
+    );
   });
 
   it("HANDLE_REDIRECT_END event does not set inProgress to None if login is in progress", (done) => {
@@ -440,17 +432,14 @@ describe("MsalBroadcastService", () => {
       }
     });
 
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.HANDLE_REDIRECT_START, InteractionType.Redirect);
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.LOGIN_START, InteractionType.Redirect);
-    // @ts-ignore
-    msalInstance.controller
-      .getEventHandler()
-      .emitEvent(EventType.HANDLE_REDIRECT_END, InteractionType.Redirect);
+    eventHandler.emitEvent(
+      EventType.HANDLE_REDIRECT_START,
+      InteractionType.Redirect
+    );
+    eventHandler.emitEvent(EventType.LOGIN_START, InteractionType.Redirect);
+    eventHandler.emitEvent(
+      EventType.HANDLE_REDIRECT_END,
+      InteractionType.Redirect
+    );
   });
 });