+++ lib/msal-common/test/client/RefreshTokenClient.spec.ts
@@ -144,6 +144,44 @@ describe("RefreshTokenClient unit tests", () => {
             config = await ClientTestUtils.createTestClientConfiguration();
         });
 
+        it("Adds correlationId to the /token query string", (done) => {
+            jest.spyOn(
+                RefreshTokenClient.prototype,
+                <any>"executePostToTokenEndpoint"
+                // @ts-expect-error
+            ).mockImplementation((url: string) => {
+                try {
+                    expect(url).toContain(
+                        `client-request-id=${TEST_CONFIG.CORRELATION_ID}`
+                    );
+                    done();
+                } catch (error) {
+                    done(error);
+                }
+            });
+
+            const client = new RefreshTokenClient(
+                config,
+                stubPerformanceClient
+            );
+            const refreshTokenRequest: CommonRefreshTokenRequest = {
+                scopes: TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
+                refreshToken: TEST_TOKENS.REFRESH_TOKEN,
+                claims: TEST_CONFIG.CLAIMS,
+                authority: TEST_CONFIG.validAuthority,
+                correlationId: TEST_CONFIG.CORRELATION_ID,
+                authenticationScheme:
+                    TEST_CONFIG.TOKEN_TYPE_BEARER as AuthenticationScheme,
+                tokenQueryParameters: {
+                    testParam: "testValue",
+                },
+            };
+
+            client.acquireToken(refreshTokenRequest).catch((e) => {
+                // Catch errors thrown after the function call this test is testing
+            });
+        });
+
         it("Adds tokenQueryParameters to the /token request", (done) => {
             jest.spyOn(
                 RefreshTokenClient.prototype,
@@ -231,8 +269,8 @@ describe("RefreshTokenClient unit tests", () => {
             );
             jest.spyOn(
                 // @ts-ignore
-                client.networkManager,
-                "sendPostRequest"
+                client.networkClient,
+                "sendPostRequestAsync"
             ).mockResolvedValue({ ...AUTHENTICATION_RESULT, headers: {} });
 
             let refreshTokenSize;
@@ -259,8 +297,8 @@ describe("RefreshTokenClient unit tests", () => {
             );
             jest.spyOn(
                 // @ts-ignore
-                client.networkManager,
-                "sendPostRequest"
+                client.networkClient,
+                "sendPostRequestAsync"
             ).mockResolvedValue({
                 ...AUTHENTICATION_RESULT_NO_REFRESH_TOKEN,
                 headers: { ...AUTHENTICATION_RESULT_WITH_HEADERS.headers },
@@ -935,8 +973,8 @@ describe("RefreshTokenClient unit tests", () => {
             const client = new RefreshTokenClient(config, performanceClient);
             jest.spyOn(
                 // @ts-ignore
-                client.networkManager,
-                "sendPostRequest"
+                client.networkClient,
+                "sendPostRequestAsync"
             ).mockResolvedValue(AUTHENTICATION_RESULT_WITH_HEADERS);
             const refreshTokenRequest: CommonRefreshTokenRequest = {
                 scopes: TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
@@ -955,6 +993,7 @@ describe("RefreshTokenClient unit tests", () => {
                     refreshTokenSize:
                         AUTHENTICATION_RESULT_WITH_HEADERS.body.refresh_token
                             .length,
+                    requestId: "xMsRequestId",
                 },
                 TEST_CONFIG.CORRELATION_ID
             );
@@ -979,8 +1018,8 @@ describe("RefreshTokenClient unit tests", () => {
             const client = new RefreshTokenClient(config, performanceClient);
             jest.spyOn(
                 // @ts-ignore
-                client.networkManager,
-                "sendPostRequest"
+                client.networkClient,
+                "sendPostRequestAsync"
             ).mockResolvedValue({ ...AUTHENTICATION_RESULT, headers: {} });
             const refreshTokenRequest: CommonRefreshTokenRequest = {
                 scopes: TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
@@ -998,6 +1037,7 @@ describe("RefreshTokenClient unit tests", () => {
                     httpVerToken: "",
                     refreshTokenSize:
                         AUTHENTICATION_RESULT.body.refresh_token.length,
+                    requestId: "",
                 },
                 TEST_CONFIG.CORRELATION_ID
             );
@@ -1446,6 +1486,53 @@ describe("RefreshTokenClient unit tests", () => {
             expect(
                 returnVal.includes(`${AADServerParamKeys.X_CLIENT_LAST_TELEM}`)
             ).toBe(false);
+        });
+    });
+
+    describe("createTokenRequestBody tests", () => {
+        it("pick up broker params", async () => {
+            const config: ClientConfiguration =
+                await ClientTestUtils.createTestClientConfiguration();
+            const client = new RefreshTokenClient(config);
+
+            const queryString =
+                // @ts-ignore
+                await client.createTokenRequestBody({
+                    scopes: ["User.Read"],
+                    redirectUri: "localhost",
+                    embeddedClientId: "child_client_id_1",
+                });
+
+            expect(queryString).toContain(`client_id=child_client_id_1`);
+            expect(queryString).toContain(
+                `brk_client_id=${config.authOptions.clientId}`
+            );
+            expect(queryString).toContain(`brk_redirect_uri=https://localhost`);
+        });
+
+        it("broker params take precedence over token body params", async () => {
+            const config: ClientConfiguration =
+                await ClientTestUtils.createTestClientConfiguration();
+            const client = new RefreshTokenClient(config);
+
+            const queryString =
+                // @ts-ignore
+                await client.createTokenRequestBody({
+                    scopes: ["User.Read"],
+                    redirectUri: "localhost",
+                    embeddedClientId: "child_client_id_1",
+                    tokenBodyParameters: {
+                        client_id: "child_client_id_2",
+                        brk_client_id: "broker_client_id_2",
+                        brk_redirect_uri: "broker_redirect_uri_2",
+                    },
+                });
+
+            expect(queryString).toContain(`client_id=child_client_id_1`);
+            expect(queryString).toContain(
+                `brk_client_id=${config.authOptions.clientId}`
+            );
+            expect(queryString).toContain(`brk_redirect_uri=https://localhost`);
         });
     });
 });