+++ samples/e2eTestUtils/src/LabConfig.ts
@@ -3,7 +3,7 @@ import { MsidApp } from "./MsidApp";
 import { MsidLab } from "./MsidLab";
 
 export type LabConfig = {
-    user: MsidUser,
-    app: MsidApp,
-    lab: MsidLab
+    user: MsidUser;
+    app: MsidApp;
+    lab: MsidLab;
 };