+++ samples/msal-angular-v3-samples/angular-b2c-sample-app/test/msa-account.spec.ts
@@ -9,6 +9,7 @@ import {
   UserTypes,
   B2cProviders,
   BrowserCacheUtils,
+  B2C_MSA_TEST_UPN,
 } from 'e2e-test-utils';
 
 const SCREENSHOT_BASE_FOLDER_NAME = `${__dirname}/screenshots/msa-account-tests`;
@@ -38,6 +39,9 @@ describe('B2C user-flow tests (msa account)', () => {
     const envResponse = await labClient.getVarsByCloudEnvironment(labApiParams);
 
     [username, accountPwd] = await setupCredentials(envResponse[0], labClient);
+
+    // TODO: Remove when B2C MSA account is available in the lab
+    username = B2C_MSA_TEST_UPN;
   });
 
   beforeEach(async () => {