+++ samples/msal-node-samples/cachePlugin.js
@@ -12,18 +12,18 @@ module.exports = function (cacheLocation) {
     const beforeCacheAccess = (cacheContext) => {
         return new Promise((resolve, reject) => {
             if (fs.existsSync(cacheLocation)) {
-                fs.readFile(cacheLocation, "utf-8", (err, data) => {
-                    if (err) {
-                        reject(err);
+                fs.readFile(cacheLocation, "utf-8", (error, data) => {
+                    if (error) {
+                        reject(error);
                     } else {
                         cacheContext.tokenCache.deserialize(data);
                         resolve();
                     }
                 });
             } else {
-               fs.writeFile(cacheLocation, cacheContext.tokenCache.serialize(), (err) => {
-                    if (err) {
-                        reject(err);
+               fs.writeFile(cacheLocation, cacheContext.tokenCache.serialize(), (error) => {
+                    if (error) {
+                        reject(error);
                     }
                     resolve();
                 });
@@ -34,9 +34,9 @@ module.exports = function (cacheLocation) {
     const afterCacheAccess = (cacheContext) => {
         return new Promise((resolve, reject) => {
             if(cacheContext.cacheHasChanged){
-                fs.writeFile(cacheLocation, cacheContext.tokenCache.serialize(), (err) => {
-                    if (err) {
-                        reject(err);
+                fs.writeFile(cacheLocation, cacheContext.tokenCache.serialize(), (error) => {
+                    if (error) {
+                        reject(error);
                     }
                     resolve();
                 });