+++ samples/msal-node-samples/username-password-cca/index.js
@@ -24,8 +24,6 @@ const usernamePasswordRequest = {
 
 cca.acquireTokenByUsernamePassword(usernamePasswordRequest).then((response) => {
     console.log("acquired token by password grant in confidential clients");
-}).catch((error) => {
-    console.log(error);
 });